Updated module augmentation for basic-styles package.
