+++ packages/ckeditor5-basic-styles/src/bold/boldediting.ts
@@ -71,9 +71,3 @@ export default class BoldEditing extends Plugin {
 		editor.keystrokes.set( 'CTRL+B', BOLD );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ BoldEditing.pluginName ]: BoldEditing;
-	}
-}
