+++ packages/ckeditor5-basic-styles/src/code/codeediting.ts
@@ -69,9 +69,3 @@ export default class CodeEditing extends Plugin {
 		inlineHighlight( editor, CODE, 'code', HIGHLIGHT_CLASS );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ CodeEditing.pluginName ]: CodeEditing;
-	}
-}
