+++ packages/ckeditor5-basic-styles/src/code/codeui.ts
@@ -60,9 +60,3 @@ export default class CodeUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ CodeUI.pluginName ]: CodeUI;
-	}
-}
