+++ packages/ckeditor5-basic-styles/src/index.ts
@@ -28,3 +28,5 @@ export { default as SuperscriptUI } from './superscript/superscriptui';
 export { default as Underline } from './underline';
 export { default as UnderlineEditing } from './underline/underlineediting';
 export { default as UnderlineUI } from './underline/underlineui';
+
+import './augmentation';
