+++ packages/ckeditor5-basic-styles/src/italic/italicediting.ts
@@ -59,9 +59,3 @@ export default class ItalicEditing extends Plugin {
 		editor.keystrokes.set( 'CTRL+I', ITALIC );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ ItalicEditing.pluginName ]: ItalicEditing;
-	}
-}
