+++ packages/ckeditor5-basic-styles/src/italic/italicui.ts
@@ -58,9 +58,3 @@ export default class ItalicUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ ItalicUI.pluginName ]: ItalicUI;
-	}
-}
