+++ packages/ckeditor5-basic-styles/src/strikethrough/strikethroughediting.ts
@@ -61,9 +61,3 @@ export default class StrikethroughEditing extends Plugin {
 		editor.keystrokes.set( 'CTRL+SHIFT+X', 'strikethrough' );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ StrikethroughEditing.pluginName ]: StrikethroughEditing;
-	}
-}
