+++ packages/ckeditor5-basic-styles/src/strikethrough/strikethroughui.ts
@@ -58,9 +58,3 @@ export default class StrikethroughUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ StrikethroughUI.pluginName ]: StrikethroughUI;
-	}
-}
