+++ packages/ckeditor5-basic-styles/src/subscript/subscriptui.ts
@@ -57,9 +57,3 @@ export default class SubscriptUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ SubscriptUI.pluginName ]: SubscriptUI;
-	}
-}
