+++ packages/ckeditor5-basic-styles/src/superscript/superscriptediting.ts
@@ -56,9 +56,3 @@ export default class SuperscriptEditing extends Plugin {
 		editor.commands.add( SUPERSCRIPT, new AttributeCommand( editor, SUPERSCRIPT ) );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ SuperscriptEditing.pluginName ]: SuperscriptEditing;
-	}
-}
