+++ packages/ckeditor5-basic-styles/src/superscript/superscriptui.ts
@@ -57,9 +57,3 @@ export default class SuperscriptUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ SuperscriptUI.pluginName ]: SuperscriptUI;
-	}
-}
