+++ packages/ckeditor5-basic-styles/src/underline/underlineediting.ts
@@ -56,9 +56,3 @@ export default class UnderlineEditing extends Plugin {
 		editor.keystrokes.set( 'CTRL+U', 'underline' );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ UnderlineEditing.pluginName ]: UnderlineEditing;
-	}
-}
