+++ packages/ckeditor5-basic-styles/src/underline/underlineui.ts
@@ -58,9 +58,3 @@ export default class UnderlineUI extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ UnderlineUI.pluginName ]: UnderlineUI;
-	}
-}
